Implement code more efficiently and update mcmod.info
Changed the implementation of the youtube downloader from a public method to a implementation that can be utilized by different files for different download types. This allows the download types to vary based on what needs to be downloaded. Also updated the MCMOD.info file so it includes the mod information and author details. Website to come after it is set up.